+++ src/PromiseInterface.php
@@ -8,4 +8,24 @@ interface PromiseInterface
      * @return PromiseInterface
      */
     public function then(callable $onFulfilled = null, callable $onRejected = null);
+
+    /**
+     * @return void
+     */
+    public function done(callable $onFulfilled = null, callable $onRejected = null);
+
+    /**
+     * @return PromiseInterface
+     */
+    public function otherwise(callable $onRejected);
+
+    /**
+     * @return PromiseInterface
+     */
+    public function always(callable $onFulfilledOrRejected);
+
+    /**
+     * @return void
+     */
+    public function cancel();
 }